api/pyproject.toml
[project]
name = "advanced-django-blueprint"
version = "0.1.0"
description = "Advanced Django Blueprint"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django>=6.0",
    "django-cors-headers>=4.9.0",
    "django-extensions>=4.1",
    "djangorestframework>=3.15.0",
    "djangorestframework-simplejwt>=5.4.0",
    "drf-spectacular[sidecar]>=0.29.0",
    "pre-commit>=4.5.1",
    "psycopg[binary]>=3.3.2",
    "python-dotenv>=1.2.1",
]

[tool.setuptools]
packages = ["app"]


[tool.ruff]
line-length = 88
target-version = "py312"
exclude = [
    ".git",
    "__pycache__",
    ".venv",
    "env",
    "venv",
    "migrations",
    "node_modules",
    "static",
    "media",
    "build",
    "dist",
    "settings",
]

[tool.ruff.lint]
select = [
    "E",    # pycodestyle errors
    "F",    # pyflakes
    "I",    # isort
    "B",    # flake8-bugbear
    "UP",   # pyupgrade
    "C90",  # mccabe complexity
]

ignore = [
    "E501",
]

fixable = ["ALL"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
line-ending = "lf"
docstring-code-format = true
docstring-code-line-length = 88
